+++ src/main/java/kr/santanrudolph/everyvent/auth/service/KakaoOAuthService.java
@@ -3,7 +3,7 @@ package kr.santanrudolph.everyvent.auth.service;
 import kr.santanrudolph.everyvent.auth.security.EveryventOAuth2User;
 import kr.santanrudolph.everyvent.domain.user.enums.SocialProvider;
 import kr.santanrudolph.everyvent.domain.user.User;
-import kr.santanrudolph.everyvent.domain.user.UserRepository;
+import kr.santanrudolph.everyvent.domain.user.repository.UserRepository;
 import kr.santanrudolph.everyvent.global.exception.ErrorCode;
 import kr.santanrudolph.everyvent.global.exception.EveryventException;
 import lombok.RequiredArgsConstructor;
